Module auth_server::token_exchanger

Utilities for exchanging authorisation codes for identity tokens and validating the resulting tokens.

For reference, see: https://developers.google.com/identity/protocols/oauth2/openid-connect#exchangecode

Structs

Claims
Claims encoded into the identity token. Reference: https://developers.google.com/identity/protocols/oauth2/openid-connect#an-id-tokens-payload

Key
A JSON Web Key potentially used for signing the identity tokens. Reference: https://auth0.com/docs/tokens/references/jwks-properties

KeySet
Container for the JSON Web Key Set containing the signing keys. Reference: https://auth0.com/docs/tokens/references/jwks-properties

TokenError
Represents a token validation error.

TokenRequest
The request paramters for the token service. Reference: https://developers.google.com/identity/protocols/oauth2/openid-connect#exchangecode

TokenResponse
The response from the token service. Reference: https://developers.google.com/identity/protocols/oauth2/openid-connect#exchangecode

Functions

decode_identity_token

exchange_code_for_token
Exchanges an authorisation code for an ID token and validates the token.

get_key
Gets the public key that matches the header's key ID and algorithm.